Create Wyam.Modules.ImageProcessor
- Move ProcessFiles and ImageProcessor modules related files to its own
module.
- Remove the dependency of ImageProcessor library and System.Drawaing
from the core.
