+++ Wyam.Core/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="ImageProcessor" version="2.2.8.0" targetFramework="net46" />
   <package id="Microsoft.CodeAnalysis.Analyzers" version="1.0.0" targetFramework="net45" />
   <package id="Microsoft.CodeAnalysis.Common" version="1.0.0" targetFramework="net45" />
   <package id="Microsoft.CodeAnalysis.CSharp" version="1.0.0" targetFramework="net45" />
