+++ Wyam.Modules.ImageProcessor/ImageFilter.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Wyam.Modules.ImageProcessor
+{
+    public enum ImageFilter
+    {
+        BlackAndWhite,
+        Comic,
+        Gotham,
+        GreyScale,
+        HiSatch,
+        Invert,
+        Lomograph,
+        LoSatch,
+        Polariod,
+        Sepia,
+    }
+}
