+++ Wyam.Modules.ImageProcessor/MetadataKeys.cs
@@ -1,0 +1,9 @@
+﻿namespace Wyam.Modules.ImageProcessor
+{
+    public static class MetadataKeys
+    {
+        public const string Base64 = "Base64";
+        public const string SourceFilePath = "SourceFilePath";
+        public const string SourceFileExt = "SourceFileExt";
+    }
+}
