+++ Wyam.Modules.ImageProcessor/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ImageProcessor" version="2.2.8.0" targetFramework="net46" />
+</packages>
